+++ components/FilterOption/FilterOption.tsx
@@ -1,22 +1,7 @@
 import { FC, useState } from 'react'
 import { IconType } from 'react-icons'
-import styled from 'styled-components'
-import { colors, fontSize } from '../styles/theme'
-
-const FilterOptionBox = styled.div`
-  border: 2px solid;
-  border-radius: 4px;
-  padding: 10px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  cursor: pointer;
-  min-width: 125px;
-`
-
-const FilterValue = styled.span`
-  font-size: ${p => p.theme.fontSize.md}px;
-`
+import { colors, fontSize } from '../../styles/theme'
+import { FilterOptionBox, FilterValue } from './styles'
 
 type Props = {
   icon: IconType
